property-underwriting-system/src/models/computer_vision/property_detector.py
from PIL import Image
from typing import List, Dict, Any
from src.api.schemas.response_models import ImageAnalysisResult

class PropertyDetector:
    async def analyze_image(self, image_path: str) -> ImageAnalysisResult:
        """Basic mock analysis - replace with real CV model later"""
        return ImageAnalysisResult(
            image_path=image_path,
            detected_objects=[{"type": "building", "confidence": 0.99}],
            damage_assessment={"damage": "none", "confidence": 0.9},
            hazard_indicators=[],
            confidence_score=0.95
        )
