property-underwriting-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
transformers==4.36.0
torch==2.1.1
torchvision==0.16.1
spacy==3.7.2
pytesseract==0.3.10
PyPDF2==3.0.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
